fix: address CodeRabbit findings on the Jobs PR
- JobsProvider.setApiClient now clears _apiClient up front, so a
  missing-credentials or getClient() failure on the new server can no
  longer leave abortJob/rerunJob/subscribeToJobs silently acting on
  the previous server's client. Adds a regression test.
- ServerJobsScreen's job list now keys each card by job id, so a
  StatefulWidget's expanded/busy state can't attach to the wrong job
  when JobsProvider's live list reorders.
- Add .toInt() after the clamp() calls in MemoryStats.freeMemory/
  appsMemory and MemorySegmentedBar's flex calculation for clarity,
  even though the current toolchain already accepts the int-typed
  clamp() result.

+++ lib/models/system_stats.dart
@@ -183,13 +183,12 @@ class MemoryStats extends Equatable {
   /// [physicalMemoryAvailable] alongside memory that is genuinely free - this
   /// is the "genuinely free" remainder once the ARC is split out.
   int get freeMemory =>
-      (physicalMemoryAvailable - arcSize).clamp(0, physicalMemoryTotal);
+      (physicalMemoryAvailable - arcSize).clamp(0, physicalMemoryTotal).toInt();
 
   /// Memory held by apps and services: everything not currently available.
-  int get appsMemory => (physicalMemoryTotal - physicalMemoryAvailable).clamp(
-    0,
-    physicalMemoryTotal,
-  );
+  int get appsMemory => (physicalMemoryTotal - physicalMemoryAvailable)
+      .clamp(0, physicalMemoryTotal)
+      .toInt();
 
   double get freeMemoryPercent {
     if (physicalMemoryTotal == 0) return 0.0;
